backend/src/requirements.txt
google.generativeai
supabase
flask
python-dotenv
pathlib
Flask-Cors
pulp
numpy